requirements.txt
deepface==0.0.79
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
retinaface==0.0.13
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0